Override session start code in test application
