+++ unit-tests/src/org/commcare/android/tests/application/AppInitializationTest.java
@@ -1,6 +1,7 @@
 package org.commcare.android.tests.application;
 
 import org.commcare.CommCareApplication;
+import org.commcare.CommCareTestApplication;
 import org.commcare.android.CommCareTestRunner;
 import org.commcare.android.util.TestAppInstaller;
 import org.commcare.dalvik.BuildConfig;
@@ -17,7 +18,7 @@ import org.robolectric.annotation.Config;
  *
  * @author Phillip Mates (pmates@dimagi.com).
  */
-@Config(application = CommCareApplication.class,
+@Config(application = CommCareTestApplication.class,
         constants = BuildConfig.class)
 @RunWith(CommCareTestRunner.class)
 public class AppInitializationTest {
