+++ unit-tests/src/org/commcare/android/tests/formnav/EndOfFormTest.java
@@ -6,12 +6,12 @@ import android.util.Log;
 import android.widget.ImageButton;
 
 import org.commcare.CommCareApplication;
+import org.commcare.CommCareTestApplication;
 import org.commcare.activities.CommCareHomeActivity;
 import org.commcare.activities.FormEntryActivity;
 import org.commcare.android.CommCareTestRunner;
 import org.commcare.android.mocks.FormAndDataSyncerFake;
 import org.commcare.android.util.TestAppInstaller;
-import org.commcare.android.util.TestUtils;
 import org.commcare.dalvik.BuildConfig;
 import org.commcare.dalvik.R;
 import org.commcare.models.AndroidSessionWrapper;
@@ -21,8 +21,6 @@ import org.commcare.session.CommCareSession;
 import org.commcare.session.SessionNavigator;
 import org.commcare.views.QuestionsView;
 import org.commcare.views.widgets.IntegerWidget;
-import org.javarosa.core.reference.ReferenceManager;
-import org.javarosa.core.reference.ResourceReferenceFactory;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -38,7 +36,7 @@ import static junit.framework.Assert.assertTrue;
 /**
  * @author Phillip Mates (pmates@dimagi.com).
  */
-@Config(application = CommCareApplication.class,
+@Config(application = CommCareTestApplication.class,
         constants = BuildConfig.class)
 @RunWith(CommCareTestRunner.class)
 public class EndOfFormTest {
