+++ unit-tests/src/org/commcare/android/tests/formsave/FormRecordProcessingTest.java
@@ -6,6 +6,7 @@ import android.util.Log;
 import android.widget.ImageButton;
 
 import org.commcare.CommCareApplication;
+import org.commcare.CommCareTestApplication;
 import org.commcare.activities.CommCareHomeActivity;
 import org.commcare.activities.FormEntryActivity;
 import org.commcare.android.CommCareTestRunner;
@@ -43,7 +44,7 @@ import static junit.framework.Assert.assertTrue;
  *
  * @author Phillip Mates (pmates@dimagi.com).
  */
-@Config(application = CommCareApplication.class,
+@Config(application = CommCareTestApplication.class,
         constants = BuildConfig.class)
 @RunWith(CommCareTestRunner.class)
 public class FormRecordProcessingTest {
